requirements.txt
dash==2.17.1
plotly==5.22.0
pyserial==3.5
pandas==2.2.2
numpy==1.26.4
